feat(headers): Add support for Retry-After header
This used to be an external crate, which seem to be abandoned and
stuck with Hyper 0.7. https://github.com/jwilm/retry-after

I think it would be better to include it with Hyper and I'm willing
to support this code here if needed. I took the code from that repo
and made sure it works with the tokio branch.